Update company controller

+++ src/main/java/com/xmnjm/repository/CompanyRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Jilion on 2017/3/5.
  */
-public interface CompanyRepository extends CrudRepository<Company, Long> {
+public interface CompanyRepository extends CrudRepository<Company, Integer> {
 }
